telegram-bot/requirements.txt
python-telegram-bot>=21.0
supabase>=2.0.0
stripe>=7.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
